+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 # RxDB Changelog
 
 <!-- CHANGELOG NEWEST -->
-
+- FIX `event-reduce` did not work when `key-compression` plugin was used.
 <!-- ADD new changes here! -->
 
 <!-- /CHANGELOG NEWEST -->